add ignoreUnusedLookups to other generate functionalities
